pyproject.toml
[project]
name = "ai-web-scraper"
version = "0.1.0"
description = "AI-powered web scraping system"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.4",
    "uvicorn[standard]==0.32.0",
    "crewai==0.83.0",
    "crewai-tools==0.17.0",
    "beautifulsoup4==4.12.3",
    "requests==2.32.3",
    "selenium==4.27.1",
    "pydantic==2.10.2",
    "python-multipart==0.0.12",
    "websockets==13.1",
    "python-dotenv==1.0.1",
    "chromadb==0.5.23",
    "langchain==0.3.7",
    "langchain-openai==0.2.8",
    "openai==1.54.4",
    "pandas==2.2.3",
    "numpy>=1.24.3,<2.0.0",
    "lxml==5.3.0",
    "aiofiles==24.1.0",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "black==24.10.0",
    "flake8==7.1.1",
    "mypy==1.13.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.hatch.build]
artifacts = [
    "src/backend/**/*"
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

